+++ tests/jasmine/unit/compiler/helpers.js
@@ -1,26 +1,30 @@
-/* global beforeEach, jasmine */
+/* global beforeEach, jasmine, addMatchersLegacy */
 
 "use strict";
 
 beforeEach(function() {
-  this.addMatchers({
-    toHaveParts: function(expected) {
-      this.message = () => "Expected " + jasmine.pp(this.actual.parts) +
-        (this.isNot ? "not " : "") +
-        " to match "+jasmine.pp(expected) + " and not be inverted, " +
-        "but it " + (this.isNot ? "did" : "didn't") + ".";
-      return this.env.equals_(this.actual.parts, expected) &&
-        this.env.equals_(this.actual.inverted, false) &&
-        this.env.equals_(this.actual.ignoreCase, false);
+  addMatchersLegacy(jasmine, {
+    toHaveParts: function(actual, expected, matchersUtil, isNot) {
+      var result = {};
+      result.message = () => "Expected " + matchersUtil.pp(actual.parts) +
+        (isNot ? "not " : "") +
+        " to match "+matchersUtil.pp(expected) + " and not be inverted, " +
+        "but it " + (isNot ? "did" : "didn't") + ".";
+      result.pass = matchersUtil.equals(actual.parts, expected) &&
+        matchersUtil.equals(actual.inverted, false) &&
+        matchersUtil.equals(actual.ignoreCase, false);
+      return result;
     },
-    toHaveInvertedParts: function(expected) {
-      this.message = () => "Expected " + jasmine.pp(this.actual.parts) +
-        (this.isNot ? "not " : "") +
-        " to match "+jasmine.pp(expected) + " and be inverted, " +
-        "but it " + (this.isNot ? "did" : "didn't") + ".";
-      return this.env.equals_(this.actual.parts, expected) &&
-        this.env.equals_(this.actual.inverted, true) &&
-        this.env.equals_(this.actual.ignoreCase, false);
+    toHaveInvertedParts: function(actual, expected, matchersUtil, isNot) {
+      var result = {};
+      result.message = () => "Expected " + matchersUtil.pp(actual.parts) +
+        (isNot ? "not " : "") +
+        " to match "+matchersUtil.pp(expected) + " and be inverted, " +
+        "but it " + (isNot ? "did" : "didn't") + ".";
+      result.pass = matchersUtil.equals(actual.parts, expected) &&
+        matchersUtil.equals(actual.inverted, true) &&
+        matchersUtil.equals(actual.ignoreCase, false);
+      return result;
     },
   });
 });