+++ HISTORY.md
@@ -1,6 +1,7 @@
 # Release History
 
 ## x.x.x (not yet released)
+* Update tests from (unsupported) jasmine-node to jasmine 6.0.0.
 
 ## 6.1.0 (2026-02-13)
 * Add 'labeled expression parameter' which allows parameterizing a rule